backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
httpx
python-jose[cryptography]
python-dotenv


